:truck: Rename services files
